decided to keep simple score timing structure without tuplets detection in this branch; will create a new branch for redesigning score to allow temporal structures with irregular rhythm patterns with tuplets
