+++ lib/musa-dsl/datasets/score-to-mxml/process-pdv.rb
@@ -41,13 +41,15 @@ module Musa::Datasets::Score::ToMXML
 
     internal_tie = false
 
-    last_duration = nil
-    last_note = nil
-
     until effective_duration_decomposition.empty?
       duration = effective_duration_decomposition.shift
 
       type, dots, tuplet_ratio = type_and_dots_and_tuplet_ratio(duration)
+
+      raise NotImplementedError,
+            "Found irregular time (tuplet ratio #{tuplet_ratio}) on element #{element}. " \
+            "Don't know how to handle on this version. " \
+        unless tuplet_ratio == 1
 
       tied = if continue_from_previous_bar && continue_to_next_bar
                'start'
@@ -71,12 +73,12 @@ module Musa::Datasets::Score::ToMXML
              end
 
       if pitch == :silence
-        last_note = measure.add_rest type: type,
+        measure.add_rest type: type,
                          dots: dots,
                          duration: (duration * divisions_per_bar).to_i,
                          voice: element[:dataset][:voice]
       else
-        last_note = measure.add_pitch pitch, octave: octave, alter: sharps,
+        measure.add_pitch pitch, octave: octave, alter: sharps,
                           type: type,
                           dots: dots,
                           duration: (duration * divisions_per_bar).to_i,
@@ -116,112 +118,6 @@ module Musa::Datasets::Score::ToMXML
 
       [p, o, s]
     end
-  end
-
-  def decompose_as_sum_of_simple_durations(duration)
-    return [] if duration.zero?
-
-    #pd = Prime.prime_division(duration.to_r.denominator).collect(&:first)
-    # TODO mejorar esta descomposición para que tenga menos factores redundantes
-    pd = Prime.prime_division(duration.to_r.denominator).collect { |base, exp| (1..exp).collect { |i| base ** i } }.flatten
-
-
-    divisors = ([[1]] + all_combinations(pd)).collect { |combination| combination.inject(:*) }
-
-    summands = []
-
-    while divisor = divisors.shift
-      c = Rational(1, divisor)
-      f = (duration / c).floor
-      n = f * c
-      summands << n unless n.zero?
-      duration -= n
-    end
-
-    raise ArgumentError, "#{duration} cannot be further decomposed" unless duration.zero?
-
-    summands
-  end
-
-  def all_combinations(numbers)
-    all_combinations = []
-    i = 1
-    until (combinations = numbers.combination(i).to_a).empty?
-      all_combinations += combinations
-      i += 1
-    end
-
-    all_combinations
-  end
-
-  def integrate_as_dotteable_durations(simple_durations)
-    integrated_durations = []
-    last = nil
-    simple_durations.each do |duration|
-      if last && duration == last / 2
-        integrated_durations[integrated_durations.size-1] += duration
-      else
-        integrated_durations << duration
-      end
-      last = duration
-    end
-    integrated_durations
-  end
-
-  def type_and_dots_and_tuplet_ratio(noteable_duration)
-    r = decompose_as_sum_of_simple_durations(noteable_duration)
-    n = r.shift
-
-    tuplet_ratio = Rational(n.denominator, nearest_lower_power_of_2(n.denominator))
-
-    type = type_of(nearest_upper_power_of_2(n))
-    dots = 0
-
-    while nn = r.shift
-      if nn == n / 2
-        dots += 1
-        n = nn
-      else
-        break
-      end
-    end
-
-    raise ArgumentError, "#{noteable_duration} cannot be decomposed as a duration with dots" unless r.empty?
-
-    return type, dots, tuplet_ratio
-  end
-
-  def nearest_upper_power_of_2(number)
-    return 0 if number.zero?
-
-    exp = Math.log2(number)
-    exp_floor = exp.floor
-    plus = exp > exp_floor ? 1 : 0
-
-    2 ** (exp_floor + plus)
-  end
-
-  def nearest_lower_power_of_2(number)
-    return 0 if number.zero?
-
-    exp_floor = Math.log2(number).floor
-
-    2 ** exp_floor
-  end
-
-  def type_of(base_type_duration)
-    duration_log2i = Math.log2(base_type_duration)
-
-    raise ArgumentError, "#{base_type_duration} is not a inverse power of 2 (i.e. 2, 1, 1/2, 1/4, 1/8, 1/64, etc)" \
-      unless base_type_duration == 2 ** duration_log2i
-
-    raise ArgumentError, "#{base_type_duration} is not between 1024th and maxima accepted durations" \
-      unless duration_log2i >= -10 && duration_log2i <= 3
-
-    ['1024th', '512th', '256th', '128th',
-     '64th', '32th', '16th', 'eighth',
-     'quarter', 'half', 'whole', 'breve',
-     'long', 'maxima'][duration_log2i + 10]
   end
 
   def midi_velocity_to_dynamics(midi_velocity)
