+++ lib/musa-dsl/datasets/score-to-mxml/process-ps.rb
@@ -5,16 +5,22 @@ module Musa::Datasets::Score::ToMXML
     case element[:dataset][:type]
     when :crescendo, :diminuendo
       if element[:change] == :start
-        measure.add_dynamics midi_velocity_to_dynamics(element[:dataset][:from]) \
+        measure.add_dynamics midi_velocity_to_dynamics(element[:dataset][:from]),
+                             placement: 'below' \
                 unless element[:dataset][:from].nil? || element[:dataset][:from] == 0
 
         measure.add_wedge element[:dataset][:type],
-                          niente: element[:dataset][:type] == :crescendo && element[:dataset][:from] == 0 ||
-                              element[:dataset][:type] == :diminuendo && element[:dataset][:to] == 0
+                          niente: element[:dataset][:type] == :crescendo && element[:dataset][:from] == 0,
+                          offset: divisions_per_bar / 4,
+                          placement: 'below'
       else
-        measure.add_wedge 'stop', offset: -divisions_per_bar / 4
+        measure.add_wedge 'stop',
+                          niente: element[:dataset][:type] == :diminuendo && element[:dataset][:to] == 0,
+                          offset: -divisions_per_bar / 4,
+                          placement: 'below'
 
-        measure.add_dynamics midi_velocity_to_dynamics(element[:dataset][:to]) \
+        measure.add_dynamics midi_velocity_to_dynamics(element[:dataset][:to]),
+                             placement: 'below' \
                 unless element[:dataset][:to].nil? || element[:dataset][:to] == 0
       end
