+++ spec/score_tomxml_1_spec.musicxml
@@ -1,0 +1,166 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<!DOCTYPE score-partwise PUBLIC "-//Recordare//DTD MusicXML 3.0 Partwise//EN" "http://www.musicxml.org/dtds/partwise.dtd">
+<score-partwise version="3.0">
+	<work>
+		<work-title>work title</work-title>
+	</work>
+	<identification>
+		<creator type="composer">Javier Sánchez Yeste</creator>
+		<encoding>
+			<encoding-date>2020-07-22</encoding-date>
+			<software>MusaDSL: MusicXML output formatter</software>
+		</encoding>
+	</identification>
+	<part-list>
+		<score-part id="piano">
+			<part-name>Piano</part-name>
+			<part-abbreviation>pno</part-abbreviation>
+		</score-part>
+	</part-list>
+	<part id="piano">
+		<measure number="1">
+			<attributes>
+				<divisions>24</divisions>
+				<time number="1">
+					<beats>4</beats>
+					<beat-type>4</beat-type>
+				</time>
+				<time number="2">
+					<beats>4</beats>
+					<beat-type>4</beat-type>
+				</time>
+				<staves>2</staves>
+				<clef number="1">
+					<sign>G</sign>
+					<line>2</line>
+				</clef>
+				<clef number="2">
+					<sign>F</sign>
+					<line>4</line>
+				</clef>
+			</attributes>
+			<direction placement="above">
+				<direction-type>
+					<metronome>
+						<beat-unit>quarter</beat-unit>
+						<per-minute>90</per-minute>
+					</metronome>
+				</direction-type>
+			</direction>
+			<direction placement="below">
+				<direction-type>
+					<dynamics>
+						<pp />
+					</dynamics>
+				</direction-type>
+			</direction>
+			<direction placement="below">
+				<direction-type>
+					<wedge type="crescendo" niente="no" />
+				</direction-type>
+				<offset sound="no">24.0</offset>
+			</direction>
+			<note>
+				<pitch>
+					<step>C</step>
+					<alter>0</alter>
+					<octave>4</octave>
+				</pitch>
+				<duration>24</duration>
+				<type>quarter</type>
+			</note>
+			<note>
+				<pitch>
+					<step>C</step>
+					<alter>0</alter>
+					<octave>4</octave>
+				</pitch>
+				<duration>24</duration>
+				<type>quarter</type>
+			</note>
+			<note>
+				<pitch>
+					<step>C</step>
+					<alter>1</alter>
+					<octave>4</octave>
+				</pitch>
+				<duration>24</duration>
+				<type>quarter</type>
+			</note>
+			<note>
+				<pitch>
+					<step>C</step>
+					<alter>0</alter>
+					<octave>4</octave>
+				</pitch>
+				<duration>24</duration>
+				<type>quarter</type>
+			</note>
+		</measure>
+		<measure number="2">
+			<note>
+				<pitch>
+					<step>D</step>
+					<alter>0</alter>
+					<octave>4</octave>
+				</pitch>
+				<duration>96</duration>
+				<type>whole</type>
+			</note>
+			<direction placement="below">
+				<direction-type>
+					<wedge type="stop" niente="no" />
+				</direction-type>
+				<offset sound="no">-24.0</offset>
+			</direction>
+			<direction placement="below">
+				<direction-type>
+					<dynamics>
+						<ff />
+					</dynamics>
+				</direction-type>
+			</direction>
+		</measure>
+		<measure number="3">
+			<direction placement="below">
+				<direction-type>
+					<wedge type="diminuendo" niente="no" />
+				</direction-type>
+				<offset sound="no">24.0</offset>
+			</direction>
+			<note>
+				<pitch>
+					<step>D</step>
+					<alter>1</alter>
+					<octave>4</octave>
+				</pitch>
+				<duration>96</duration>
+				<type>whole</type>
+			</note>
+		</measure>
+		<measure number="4">
+			<note>
+				<pitch>
+					<step>E</step>
+					<alter>0</alter>
+					<octave>4</octave>
+				</pitch>
+				<duration>96</duration>
+				<type>whole</type>
+			</note>
+			<direction placement="below">
+				<direction-type>
+					<wedge type="stop" niente="no" />
+				</direction-type>
+				<offset sound="no">-24.0</offset>
+			</direction>
+			<direction placement="below">
+				<direction-type>
+					<dynamics>
+						<pp />
+					</dynamics>
+				</direction-type>
+			</direction>
+		</measure>
+	</part>
+</score-partwise>
